refactor(ios): resolve auth diagnostics review findings
Keep Supabase mapping out of the auth types file and align the startup invariant test with the project naming convention.
